gtoContainer.py: Explicitly set property size & width when reading data
In the Python interface to the gto file access there was no place to set
the Python object's representation of each properties' size and width.
This caused a problem, as Property.setData warns, that kept the
container from successfully writing because the property contents don't
match their advertised size. This change allows reading and writing.